Removed some warnings (e.g. in CCP4VolumeData, CoarseGrainDataLoader etc). Started CUDA version of contour buildup algorithm.
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/plugins/protein_cuda/trunk@290 8215d0b1-3b2f-4a13-8d91-4b3e73067666

+++ CCP4VolumeData.cpp
@@ -128,7 +128,7 @@ bool protein::CCP4VolumeData::tryLoadFile(void) {
 
     // get length of file:
     fin.seekg( 0, std::ios::end);
-    unsigned int fileLength = fin.tellg();
+    unsigned int fileLength = static_cast<unsigned int>(fin.tellg());
     fin.seekg( 0, std::ios::beg);
 
 #define WORD 4
